requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
requests>=2.32.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
pandas>=2.2.0
langchain-core>=0.3.0
gunicorn>=22.0.0
pytest>=8.0.0
pytest-flask>=1.3.0
# psycopg2-binary  # Uncomment for PostgreSQL (requires C build tools on Windows)
